fix(hub-route): unique 제약 위반만 중복으로 매핑하도록 개선

+++ services/logistics-service/src/main/java/com/sparta/gt5lt7/logisticsservice/domain/entity/HubRoute.java
@@ -35,12 +35,12 @@ public class HubRoute extends BaseEntity {
     private UUID toHubId;
 
     @Column(name = "distance", nullable = false)
-    private Double distance;
+    private Integer distance;
 
     @Column(name = "duration", nullable = false)
-    private Double duration;
+    private Integer duration;
 
-    public static HubRoute create(UUID fromHubId, UUID toHubId, Double distance, Double duration) {
+    public static HubRoute create(UUID fromHubId, UUID toHubId, Integer distance, Integer duration) {
         return HubRoute.builder()
                 .fromHubId(fromHubId)
                 .toHubId(toHubId)
